fix: 修复报错；

+++ lib/src/plugin/heat_map_tiles_provider.dart
@@ -81,16 +81,17 @@ class HeatMapTilesProvider extends TileProvider {
         var cell = grid[y][x];
 
         if (cell == null) {
-          grid[y][x] = DataPoint(pixel.x, pixel.y, k);
+          grid[y][x] = DataPoint(pixel.x.toDouble(), pixel.y.toDouble(), k);
           cell = grid[y][x];
         } else {
-          cell.merge(pixel.x, pixel.y, k);
+          cell.merge(pixel.x.toDouble(), pixel.y.toDouble(), k);
         }
         localMax = math.max(cell!.z, localMax);
         localMin = math.min(cell.z, localMin);
 
         if (bounds.contains(point.latLng)) {
-          filteredData.add(DataPoint(pixel.x, pixel.y, k));
+          filteredData
+              .add(DataPoint(pixel.x.toDouble(), pixel.y.toDouble(), k));
         }
       }
     }
